homepage update - [mobile-client]
